+++ src/test/java/com/echobox/api/linkedin/jsonmapper/LinkedInJsonMapperV1Test.java
@@ -23,8 +23,8 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
-import com.echobox.api.linkedin.types.Company;
-import com.echobox.api.linkedin.types.Location;
+import com.echobox.api.linkedin.types.v1.Company;
+import com.echobox.api.linkedin.types.v1.Location;
 
 import org.junit.Test;
